When you try your best but you don't succeed
Tried adding fading of colours for circle with variables with if counter statement. Works but sort of doesn't and not in the way I want it to. A bit difficult to properly figure out how to make it like what I want especially because my variables and everything is so confusing and all over the place.